requirements.txt
aiodns==3.1.1
aiohttp==3.8.2
aiosignal==1.3.1
alpaca-py==0.10.0
alpaca-trade-api==3.1.1
alpha-vantage==2.3.1
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
ccxt==3.0.61
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.1.1
exchange_calendars==4.5.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.50.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.3.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
ibapi==9.81.1.post1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpickle==3.0.3
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lumibot==2.8.11
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
more-itertools==10.2.0
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.0.3
multidict==5.2.0
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.0.3
pandas-datareader==0.10.0
pandas_market_calendars==4.4.0
pandas_ta==0.3.14b0
parso==0.8.3
passlib==1.7.4
peewee==3.17.1
pillow==10.2.0
plotly==5.20.0
pluggy==1.4.0
polygon-api-client==1.13.4
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycares==4.4.0
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
pyluach==2.2.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
python-telegram-bot==21.0.1
pytz==2024.1
PyYAML==6.0.1
Quandl==3.7.0
QuantStats==0.0.62
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.28
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
timedelta==2020.12.3
tk==0.1.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
yfinance==0.2.37
